third_party/WebKit/LayoutTests/platform/linux/http/tests/serviceworker/navigation-preload/chromium/navigation-preload-resource-timing-expected.txt

This is a testharness.js-based test.
FAIL Navigation Preload Resource Timing after terminating SW. assert_equals: The second result must contain two entry. expected 2 but got 1
Harness: the test ran to completion.
